Initial work for python3 compatibility, currently fails many tests
